requirements.txt
djangorestframework==3.5.3
Markdown==2.6.7
django-filter==1.0.0
django==1.9.7
pygments
